refactor tractor 1.0 moved helper functions down and changed where some state lives
