+++ lib/components/Application.js
@@ -1,17 +1,16 @@
-import React, { Component } from 'react'
+import React, { Component } from 'react';
 import firebase, { reference, signIn } from '../firebase';
-import { pick, map, extend, keyBy } from 'lodash';
+import { pick, map, extend } from 'lodash';
 import Title from './Title';
 import ChatWindow from './ChatWindow';
 import MessageBar from './MessageBar';
-import moment from 'moment';
 
 export default class Application extends Component {
-
   constructor() {
     super();
     this.state = {
       messages: [],
+      filteredMessages: [],
       draftMessage: '',
       user: null,
     };
@@ -31,39 +30,14 @@ export default class Application extends Component {
     const { user, messages, draftMessage } = this.state;
     return (
       <div className="Application">
-        <Title
-          reversing={this.reversing.bind(this)}
-        />
+        <Title reversing={this.reversing.bind(this)} />
         <ChatWindow
           messages={messages}
-          getUserArray={this.getUserArray.bind(this)}
-          searchUserArray={this.searchUserArray.bind(this)}
-        />
-        <MessageBar
-          draftMessage={draftMessage}
-          user={user}
-          addNewMessage={this.addNewMessage.bind(this)}
-          handleChange = {this.handleChange.bind(this)}
-          clearField={this.clearField.bind(this)}
-        />
+          searchUserArray={this.searchUserArray.bind(this)} />
+        <MessageBar user={user} />
       </div>
     )
   }
-
-  addNewMessage() {
-    const now = moment().format("MMMM Do, h:mma");
-    const { user, draftMessage } = this.state;
-    reference.push({
-      user: pick(user, 'displayName', 'email', 'uid'),
-      content: draftMessage,
-      createdAt: now,
-    });
-    this.clearField();
-  }
-
-  handleChange(e) {this.setState({ draftMessage: e.target.value })};
-
-  clearField() {this.setState({ draftMessage: '' })}
 
   reversing() {
     const messageArray = this.state.messages;
@@ -77,17 +51,6 @@ export default class Application extends Component {
       return (message.user.uid === userID);
     })
     this.setState({messages: userMessageArray});
-  }
-
-  getUserArray() {
-    let userList = keyBy(this.state.messages, 'user.displayName');
-    let userNameArray = (Object.keys(userList));
-    let userArray = [];
-    userNameArray.forEach((userName)=>{
-      userArray.push(userList[userName].user);
-    });
-    console.log(userArray)
-    return userArray;
   }
 
 }
