+++ lib/components/ChatWindow.js
@@ -9,9 +9,8 @@ export default class ChatWindow extends React.Component {
       <div className = "main-window">
         <Chat messages={this.props.messages} />
         <Users
-          getUserArray={this.props.getUserArray}
-          searchUserArray={this.props.searchUserArray}
-        />
+          messages={this.props.messages}
+          searchUserArray={this.props.searchUserArray} />
       </div>
     )
   }
