+++ lib/components/MessageBar.js
@@ -1,12 +1,14 @@
 import React from 'react';
 import firebase, { reference, signIn } from '../firebase';
 import { pick } from 'lodash';
+import moment from 'moment';
 
 export default class MessageBar extends React.Component {
   constructor() {
     super();
     this.state = {
       length: 0,
+      draftMessage: '',
     }
   }
 
@@ -22,16 +24,16 @@ export default class MessageBar extends React.Component {
         <div className="type-here">
           <input type="text"
             placeholder="Message"
-            value={this.props.draftMessage}
-            onChange={ (e) => this.props.handleChange(e)}
+            value={this.state.draftMessage}
+            onChange={ (e) => this.handleChange(e)}
             onKeyUp={ () => this.charCount() }
           />
           <span id="count">{this.state.length}</span>
           <button className="submit"
             disabled={this.state.length > 140}
-            onClick={ () => { this.props.addNewMessage(); this.setState({ length: 0 }); }}>
+            onClick={ () => { this.addNewMessage(); this.setState({ length: 0 }); }}>
             Submit</button>
-          <button className="clear" onClick={ () => this.props.clearField() }>
+          <button className="clear" onClick={ () => this.clearField() }>
             Clear</button>
         </div>
     </div>
@@ -39,9 +41,23 @@ export default class MessageBar extends React.Component {
   }
 
   charCount() {
-    const length = this.props.draftMessage.length;
+    const length = this.state.draftMessage.length;
     const max = 140;
     this.setState({ length: length });
   }
 
+  addNewMessage() {
+    const now = moment().format("MMMM Do, h:mma");
+    const { user, draftMessage } = this.props;
+    reference.push({
+      user: pick(user, 'displayName', 'email', 'uid'),
+      content: this.state.draftMessage,
+      createdAt: now,
+    });
+    this.clearField();
+  }
+
+  handleChange(e) {this.setState({ draftMessage: e.target.value })};
+
+  clearField() {this.setState({ draftMessage: '' })}
 }
