+++ lib/components/users.js
@@ -1,8 +1,11 @@
 import React from 'react';
+import {keyBy} from 'lodash';
 
 export default class Users extends React.Component {
+
   render() {
-    const userArray = this.props.getUserArray();
+    const userArray = this.getUserArray();
+
     return(
       <div>
         <ul id = "users">
@@ -11,4 +14,15 @@ export default class Users extends React.Component {
       </div>
     )
   }
+
+  getUserArray() {
+    let userList = keyBy(this.props.messages, 'user.displayName');
+    let userNameArray = (Object.keys(userList));
+    let userArray = [];
+    userNameArray.forEach((userName)=>{
+      userArray.push(userList[userName].user);
+    });
+    return userArray;
+  }
+  
 }
